+++ src/main/java/com/codeit/weatherwear/global/event/listener/NotificationKafkaListener.java
@@ -24,7 +24,7 @@ import org.springframework.stereotype.Component;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class NotificationEventListener {
+public class NotificationKafkaListener {
 
   private final NotificationService notificationService;
