+++ src/main/java/com/codeit/weatherwear/global/event/publisher/SseKafkaPublisher.java
@@ -1,4 +1,4 @@
-package com.codeit.weatherwear.global.event.listener;
+package com.codeit.weatherwear.global.event.publisher;
 
 import com.codeit.weatherwear.domain.notification.dto.NotificationDto;
 import com.codeit.weatherwear.global.event.dto.MultipleNotificationCreatedEvent;
@@ -6,13 +6,11 @@ import com.codeit.weatherwear.global.event.dto.NotificationCreatedEvent;
 import com.codeit.weatherwear.global.event.exception.KafkaMessageConvertException;
 import com.codeit.weatherwear.global.properties.KafkaTopics;
 import com.codeit.weatherwear.global.sse.SseMessage;
-import com.codeit.weatherwear.global.sse.SseService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.concurrent.CompletableFuture;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.kafka.core.KafkaTemplate;
 import org.springframework.kafka.support.SendResult;
 import org.springframework.scheduling.annotation.Async;
@@ -23,10 +21,9 @@ import org.springframework.transaction.event.TransactionalEventListener;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class SseEventListener {
+public class SseKafkaPublisher {
 
   private final KafkaTemplate<String, String> kafkaTemplate;
-  private final SseService sseService;
   private final ObjectMapper objectMapper;
   private final KafkaTopics kafkaTopics;
 
@@ -44,21 +41,6 @@ public class SseEventListener {
     event.notificationDtos().stream()
         .map(notificationDto -> SseMessage.create(notificationDto.receiverId(), notificationDto))
         .forEach(this::sendToKafka);
-  }
-
-  @Async("eventExecutor")
-  @KafkaListener(
-      topics = "${spring.kafka.topics.sse-sent}",
-      groupId = "#{T(java.util.UUID).randomUUID().toString()}"
-  )
-  public void handleSseSendEvent(String kafkaEvent) {
-    SseMessage sseMessage = null;
-    try {
-      sseMessage = objectMapper.readValue(kafkaEvent, SseMessage.class);
-    } catch (JsonProcessingException e) {
-      throw KafkaMessageConvertException.withEvent(kafkaEvent);
-    }
-    sseService.send(sseMessage);
   }
 
   private void sendToKafka(SseMessage sseMessage) {
@@ -80,4 +62,5 @@ public class SseEventListener {
       throw KafkaMessageConvertException.withEvent(sseMessage.toString());
     }
   }
+
 }
